+++ CriteoPublisherSdk/Sources/Standalone/CRInterstitial.h
@@ -36,6 +36,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)loadAd;
 - (void)loadAdWithBidToken:(CRBidToken *)bidToken;
+- (void)loadAdWithDisplayData:(NSString *)displayData;
 
 - (void)presentFromRootViewController:(UIViewController *)rootViewController;