+++ CriteoPublisherSdk/Tests/UnitTests/Standalone/CRBannerViewTests.m
@@ -102,7 +102,7 @@
 - (void)testWithRendering {
   WKWebView *realWebView = [WKWebView new];
 
-  CR_CdbBid *bid = [self cdbBidWithDisplayUrl:@""];
+  CR_CdbBid *bid = [self cdbBidWithDisplayUrl:@"-"];
   OCMStub([self.criteo getBid:[self expectedCacheAdUnit]]).andReturn(bid);
 
   CRBannerView *bannerView = [self bannerViewWithWebView:realWebView];
@@ -248,9 +248,23 @@
   [self cr_waitShortlyForExpectations:@[ openInBrowserExpectation ]];
 }
 
-#pragma inhouseSpecificTests
+- (void)testDisplayAdWithDataSuccess {
+  MockWKWebView *mockWebView = [MockWKWebView new];
+  NSString *displayURL =
+      @"https://rdi.eu.criteo.com/delivery/r/ajs.php?did=5c98e9d9c574a3589f8e9465fce67b00&u=%7Cx8O2jgV2RMISbZvm2b09FrpmynuoN27jeqtp1aMfZdU%3D%7C&c1=oP5_e7JVVt0EkjVehxP6aIOIWS-fm2fzhyMXUboeuR1zkGydE3HlloxT1QAbHNNgeH7t9e1IR6mv0biMxm46ZSFdAXZXreJVeP6QwU8IPLUsA32HNafhqgpnKTwmx9RrrJm4CS5Wqj07vNY7UTgDei8AWqc5CGPT2wm7W02JRvgN2kA-oWbWifmmm6EPpqVZijDHDzXwaNgzrfsaEodEmYAjFepGF0mdElHoFUCPKuOtc7mUQijLG0BSS9RhwrCTcAv42KkEQ359Et_eDnQcSt9OAF3bL64QIvLQxt2ekYFNuv3zng03qL0DIHS2bDJwRb3ieUlvZCWHI49OqM5PqoGDpSzdhdwfTE18L6cOOVKqPQ0dPofN4dkSs9IbVGiYlPnjfibL88PwTspYvki2svidSDIa2agQMHVgEof8YY4x4VgPjA8XY-s93ttw_i-RN3lcQn2mGEp6FYmRsyjFEDxHgGfJ0j6U";
 
-- (void)testLoadingSuccess {
+  CRBannerView *bannerView = [self bannerViewWithWebView:mockWebView];
+  [bannerView loadAdWithDisplayData:displayURL];
+
+  XCTAssertTrue([mockWebView.loadedHTMLString
+      containsString:
+          @"<script src=\"https://rdi.eu.criteo.com/delivery/r/ajs.php?did=5c98e9d9c574a3589f8e9465fce67b00&u=%7Cx8O2jgV2RMISbZvm2b09FrpmynuoN27jeqtp1aMfZdU%3D%7C&c1=oP5_e7JVVt0EkjVehxP6aIOIWS-fm2fzhyMXUboeuR1zkGydE3HlloxT1QAbHNNgeH7t9e1IR6mv0biMxm46ZSFdAXZXreJVeP6QwU8IPLUsA32HNafhqgpnKTwmx9RrrJm4CS5Wqj07vNY7UTgDei8AWqc5CGPT2wm7W02JRvgN2kA-oWbWifmmm6EPpqVZijDHDzXwaNgzrfsaEodEmYAjFepGF0mdElHoFUCPKuOtc7mUQijLG0BSS9RhwrCTcAv42KkEQ359Et_eDnQcSt9OAF3bL64QIvLQxt2ekYFNuv3zng03qL0DIHS2bDJwRb3ieUlvZCWHI49OqM5PqoGDpSzdhdwfTE18L6cOOVKqPQ0dPofN4dkSs9IbVGiYlPnjfibL88PwTspYvki2svidSDIa2agQMHVgEof8YY4x4VgPjA8XY-s93ttw_i-RN3lcQn2mGEp6FYmRsyjFEDxHgGfJ0j6U\"></script>"]);
+  XCTAssertEqualObjects([NSURL URLWithString:@"https://criteo.com"], mockWebView.loadedBaseURL);
+}
+
+#pragma - In House
+
+- (void)testLoadingWithTokenSuccess {
   MockWKWebView *mockWebView = [MockWKWebView new];
   CRBidToken *token = [[CRBidToken alloc] initWithUUID:[NSUUID UUID]];
   NSString *displayURL =